firmware: Reorganize guest and linux/guest modules
