+++ src/firmware/linux/guest/mod.rs
@@ -1,3 +1,4 @@
 // SPDX-License-Identifier: Apache-2.0
+
 pub(crate) mod ioctl;
 pub(crate) mod types;
